fix: use pairs instead of ipairs to iterate over tabpages (#56)
Using Lua's ipairs() to iterate over tabpages is incorrect because
tabpage indices are not guaranteed to be sequential nor start from 1.
Lua's ipairs() stops iterating over integer keys as soon as it
encounters a nil value, which can lead to some tabpages not being
destroyed. Using pairs() ensures that all tabpages are destroyed.